+++ app/domain/repository/status.go
@@ -10,7 +10,7 @@ import (
 
 type Status interface {
 	// Fetch statuses which has specified id
-	FindByID(ctx context.Context, id int) (*object.Status, error)
+	FindByID(ctx context.Context, id int64) (*object.Status, error)
 	// Create a new status
 	Create(ctx context.Context, tx *sqlx.Tx, status *object.Status) error
 }
